+++ packages/connectivity/src/scp-cf/destination/destination-accessor-proxy-configuration.spec.ts
@@ -48,8 +48,9 @@ describe('proxy configuration', () => {
     ];
     process.env['https_proxy'] = 'some.proxy.com:1234';
 
-    const actual = await getDestination(destinationName, {
-      userJwt: subscriberServiceToken,
+    const actual = await getDestination({
+      destinationName,
+      jwt: subscriberServiceToken,
       cacheVerificationKeys: false,
       iasToXsuaaTokenExchange: false
     });
@@ -86,8 +87,9 @@ describe('proxy configuration', () => {
       }
     };
 
-    const actual = await getDestination('OnPremise', {
-      userJwt: subscriberServiceToken,
+    const actual = await getDestination({
+      destinationName: 'OnPremise',
+      jwt: subscriberServiceToken,
       cacheVerificationKeys: false,
       iasToXsuaaTokenExchange: false
     });
@@ -124,7 +126,7 @@ describe('proxy configuration', () => {
         }
       }
     };
-    const actual = await getDestination('OnPremise');
+    const actual = await getDestination({ destinationName: 'OnPremise' });
     expect(actual).toEqual(expected);
     httpMocks.forEach(mock => expect(mock.isDone()).toBe(true));
   });